Progress on homework 4

+++ src/Homework4/Item.java
@@ -2,21 +2,22 @@ package Homework4;
 
 public class Item {
     private String name;
-    private double price;
+    private int price;
 
-    public Item(String name, double price) {
+    // Item Constructor
+    public Item(String name, int price) {
         this.name = name;
         this.price = price;
     }
-
+    // returns name of Item
     public String getName() {
         return name;
     }
-
-    public double getPrice() {
+    // returns individual price of item
+    public int getPrice() {
         return price;
     }
-
+    // returns string representation of item
     public String toString() {
         return "name: " + name + "price: " + price;
     }
